simpler & more customizable menu builder, updated docs, added example
