+++ example/html/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Electron Context Menu - Example</title>
+</head>
+<body>
+  <div id="content">
+    <h1>Electron Context Menu - Example</h1>
+    <p>test index page</p>
+    <textarea name="textarea" id="" cols="30" rows="10"></textarea>
+  </div>
+  
+  <script>
+    console.log('hello from index')
+  </script>
+</body>
+</html>
