+++ lib/index.js
@@ -1,72 +1,160 @@
 "use strict";
 
-require("core-js/modules/es6.object.define-property");
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.buildMenus = void 0;
-
 require("core-js/modules/es6.array.map");
 
-require("core-js/modules/es6.regexp.to-string");
+var _lodash = _interopRequireDefault(require("lodash.clonedeep"));
 
-require("core-js/modules/es6.date.to-string");
+var _lodash2 = _interopRequireDefault(require("lodash.defaults"));
 
-require("core-js/modules/es6.object.to-string");
+var _lodash3 = _interopRequireDefault(require("lodash.isempty"));
 
-var _menuTemplates = require("./menuTemplates");
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-var _buildContextMenu = require("./buildContextMenu");
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-var buildMenus = function buildMenus(editable, spellChecker) {
-  var ctxMenu;
+var remote = require('electron').remote;
 
-  if (!editable) {
-    // if right click in uneditable area
-    if (window.getSelection().toString() === '') {
-      // if no text is selected
-      ctxMenu = (0, _buildContextMenu.BuildContextMenu)(_menuTemplates.templates.reload);
-    } else {
-      // if text is selected
-      ctxMenu = (0, _buildContextMenu.BuildContextMenu)(_menuTemplates.templates.copy);
+var Menu = remote.require('electron').Menu; // defining default menu templates **********************************
+
+
+var copyMenuTemplate = [{
+  role: 'copy'
+}];
+var pasteMenuTemplate = [{
+  role: 'paste'
+}, {
+  type: 'separator'
+}, {
+  role: 'selectall'
+}];
+var reloadMenuTemplate = [{
+  role: 'reload'
+}];
+var editorMenuTemplate = [{
+  role: 'cut'
+}, {
+  role: 'copy'
+}, {
+  role: 'paste'
+}, {
+  type: 'separator'
+}, {
+  role: 'selectall'
+}]; // template when spell check can't find suggestions
+
+var noSuggestionsTemplate = {
+  exists: true,
+  menuItems: [{
+    label: 'No Suggestions',
+    enabled: false
+  }] // individual suggestion menu item prototype
+
+};
+
+var SuggestionMenuItem = function SuggestionMenuItem(correction) {
+  _classCallCheck(this, SuggestionMenuItem);
+
+  if (typeof correction !== 'string') {
+    throw new Error('SuggestionMenuItem: parameter must be a string');
+  }
+
+  this.item = {
+    label: correction,
+    click: function click() {
+      remote.getCurrentWebContents().replaceMisspelling(correction);
     }
-  } else {
-    // if right click in editable textarea
-    if (window.getSelection().toString() === '') {
-      // no text is selected in editable textarea
-      ctxMenu = (0, _buildContextMenu.BuildContextMenu)(_menuTemplates.templates.paste);
-    } else {
-      // some text is selected in editable textarea
-      var selection = window.getSelection().toString();
+  };
+}; // prefix or suffix menu template
 
-      if (!spellChecker.isMisspelled(selection)) {
-        // selected word is spelled correctly
-        ctxMenu = (0, _buildContextMenu.BuildContextMenu)(_menuTemplates.templates.editor);
-      } else {
-        // selected word is misspelled
-        var suggestions = spellChecker.getCorrectionsForMisspelling(selection);
 
-        if (suggestions.length === 0) {
-          // no suggestions foing in dictionary
-          ctxMenu = (0, _buildContextMenu.BuildContextMenu)(_menuTemplates.templates.editor, _menuTemplates.templates.noSuggest);
-        } else {
-          // suggestions were found
-          suggestions = suggestions.slice(0, 3); // build list of suggestion menu items with their click functions
+var MenuTemplate = function MenuTemplate(list) {
+  _classCallCheck(this, MenuTemplate);
 
-          var suggestedItemList = suggestions.map(function (suggestion) {
-            var menuItem = new _menuTemplates.templates.SuggestionMenuItem(suggestion);
-            return menuItem.item;
-          }); // build suggestions template object and pass to BuildContextMenu as prefix
+  this.menu = {
+    exists: true,
+    menuItems: list
+  };
+};
 
-          var suggestedMenu = new _menuTemplates.templates.SuggestionsMenuTemplate(suggestedItemList);
-          ctxMenu = (0, _buildContextMenu.BuildContextMenu)(_menuTemplates.templates.editor, suggestedMenu.menu);
-        }
-      }
+var menuTemplates = {
+  copy: copyMenuTemplate,
+  editor: editorMenuTemplate,
+  noSuggest: noSuggestionsTemplate,
+  paste: pasteMenuTemplate,
+  reload: reloadMenuTemplate,
+  SuggestionMenuItem: SuggestionMenuItem,
+  MenuTemplate: MenuTemplate
+};
+
+var BuildContextMenu = function BuildContextMenu() {
+  var menuTemplate = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : editorMenuTemplate;
+  var prefix = arguments.length > 1 ? arguments[1] : undefined;
+  var suffix = arguments.length > 2 ? arguments[2] : undefined;
+
+  /*
+    builds context menu in the clicked area in browserwindow of electron app
+     Args:
+      menuTemplate: Object - contains template for basic underlying menu
+      prefix: Object - contains any prefix options required in the ctx menu
+      suffix: Object - contains and suffix options required in the ctx menu
+    Returns:
+      Menu: Electron Menu
+    Usage:
+      for just simple menus no prefis or suffix in either menus
+        buildContextMenu({}, {})
+      for menus with prefixs and or suffixes
+        buildContextMenu({}, suffix) - no prefix, only suffix
+  */
+  // deepClone makes sure the object is completely cleared and prevents
+  // redundent or repeated copies of menu items
+  var template = (0, _lodash["default"])(menuTemplate); // defaults for prefix and suffix menu objects
+  // in case no suffix or prefix items are passed into constructor
+
+  prefix = (0, _lodash2["default"])({}, prefix, {
+    exists: false,
+    menuItems: []
+  });
+  suffix = (0, _lodash2["default"])({}, suffix, {
+    exists: false,
+    menuItems: []
+  }); // attach prefix items to menu template
+
+  if (prefix.exists) {
+    var prefixMenu = prefix.menuItems;
+
+    if (!(0, _lodash3["default"])(prefixMenu)) {
+      template.unshift.apply(template, prefixMenu.map(function (item) {
+        return {
+          label: item.label,
+          click: item.click
+        };
+      }).concat({
+        type: 'separator'
+      }));
+    }
+  } // attach suffix items to menu template
+
+
+  if (suffix.exists) {
+    var suffixMenu = suffix.menuItems;
+
+    if (!(0, _lodash3["default"])(suffixMenu)) {
+      template.push({
+        type: 'separator'
+      });
+      template.push.apply(template, suffixMenu.map(function (item) {
+        return {
+          label: item.label,
+          click: item.click
+        };
+      }));
     }
   }
 
-  return ctxMenu;
+  return Menu.buildFromTemplate(template);
 };
 
-exports.buildMenus = buildMenus;
+module.exports = {
+  BuildContextMenu: BuildContextMenu,
+  menuTemplates: menuTemplates
+};
